+++ whatsnew.md
@@ -1,4 +1,4 @@
 # What's new
 
 - UI refinements
-- Control number of app icon columns on small and medium screens
+- Control number of app icon columns on small, medium and large screens